added constants package for all constants
Took all the (inner) classes in BE consts + constants.py and split/put them in dedicated .py files, one for each class (except for inner classes in BE consts)
