+++ anyway/accidents_around_schools.py
@@ -5,7 +5,7 @@ import pandas as pd
 import sqlalchemy as sa
 from sqlalchemy import or_
 
-from anyway.backend_constants import BE_CONST
+from anyway.constants.backend_constants import BackEndConstants
 from anyway.models import AccidentMarker, Involved, School
 from anyway.app_and_db import db
 
@@ -57,8 +57,8 @@ def acc_inv_query(longitude, latitude, distance, start_date, end_date, school):
         .filter(AccidentMarker.provider_and_id == Involved.provider_and_id)
         .filter(
             or_(
-                (AccidentMarker.provider_code == BE_CONST.CBS_ACCIDENT_TYPE_1_CODE),
-                (AccidentMarker.provider_code == BE_CONST.CBS_ACCIDENT_TYPE_3_CODE),
+                (AccidentMarker.provider_code == BackEndConstants.CBS_ACCIDENT_TYPE_1_CODE),
+                (AccidentMarker.provider_code == BackEndConstants.CBS_ACCIDENT_TYPE_3_CODE),
             )
         )
         .filter(AccidentMarker.created >= start_date)
