+++ anyway/widgets/suburban_widgets/accident_count_by_accident_year_widget.py
@@ -3,7 +3,7 @@ from typing import Dict
 from flask_babel import _
 
 from anyway.request_params import RequestParams
-from anyway.backend_constants import AccidentSeverity
+from anyway.constants.accident_severity import AccidentSeverity
 from anyway.infographics_dictionaries import segment_dictionary
 from anyway.models import AccidentMarkerView
 from anyway.widgets.widget import register
