+++ anyway/widgets/suburban_widgets/accident_count_by_driver_type_widget.py
@@ -5,7 +5,7 @@ from typing import Dict
 from flask_babel import _
 
 from anyway.request_params import RequestParams
-from anyway.backend_constants import DriverType
+from anyway.constants.driver_type import DriverType
 from anyway.widgets.widget_utils import get_accidents_stats, get_injured_filters
 from anyway.models import InvolvedMarkerView
 from anyway.vehicle_type import VehicleCategory
