+++ anyway/widgets/suburban_widgets/top_road_segments_accidents_per_km_widget.py
@@ -7,7 +7,7 @@ from sqlalchemy import cast, func, Numeric, desc
 # noinspection PyProtectedMember
 from flask_babel import _
 from anyway.request_params import RequestParams
-from anyway.backend_constants import AccidentSeverity
+from anyway.constants.accident_severity import AccidentSeverity
 from anyway.widgets.widget import register
 from anyway.widgets.widget_utils import get_query
 from anyway.models import AccidentMarkerView
