+++ translations/he/LC_MESSAGES/messages.po
@@ -7,7 +7,7 @@ msgid ""
 msgstr ""
 "Project-Id-Version: PROJECT VERSION\n"
 "Report-Msgid-Bugs-To: EMAIL@ADDRESS\n"
-"POT-Creation-Date: 2022-01-08 14:56+0200\n"
+"POT-Creation-Date: 2022-01-15 12:35+0200\n"
 "PO-Revision-Date: 2020-10-16 15:42+0000\n"
 "Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
 "Language: he\n"
@@ -16,19 +16,7 @@ msgstr ""
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
-"Generated-By: Babel 2.9.1\n"
-
-#: anyway/backend_constants.py:149
-msgid "killed"
-msgstr "הרוג/ה"
-
-#: anyway/backend_constants.py:150
-msgid "severe injured"
-msgstr "פצוע/ה קשה"
-
-#: anyway/backend_constants.py:151
-msgid "light injured"
-msgstr "פצוע/ה קל"
+"Generated-By: Babel 2.6.0\n"
 
 #: anyway/flask_app.py:677 anyway/flask_app.py:782
 msgid "Discussion not found:"
@@ -262,6 +250,18 @@ msgstr "אופניים/קורקינט"
 msgid "other"
 msgstr "אחרות"
 
+#: anyway/constants/backend_constants.py:109
+msgid "killed"
+msgstr "הרוג/ה"
+
+#: anyway/constants/backend_constants.py:110
+msgid "severe injured"
+msgstr "פצוע/ה קשה"
+
+#: anyway/constants/backend_constants.py:111
+msgid "light injured"
+msgstr "פצוע/ה קל"
+
 #: anyway/widgets/suburban_widgets/accident_count_by_accident_type_widget.py:46
 msgid "Number of accidents by accident type"
 msgstr "מספר תאונות לפי סוג"
@@ -278,7 +278,7 @@ msgstr ""
 "כמות התאונות הקטלניות, הקשות והקלות במקטע, לפי שנת התאונה, מחולק לפי "
 "חומרת התאונה, בפרק הזמן הנבחר."
 
-#: anyway/widgets/suburban_widgets/accident_count_by_car_type_widget.py:114
+#: anyway/widgets/suburban_widgets/accident_count_by_car_type_widget.py:115
 msgid ""
 "comparing vehicle type percentage in accidents in {} relative to national"
 " average"
@@ -322,27 +322,27 @@ msgstr "מספר תאונות לפי חומרה"
 msgid "Fatal, severe and light accidents count in the specified location."
 msgstr "כמות התאונות הקטלניות, הקשות והקלות במקטע בפרק הזמן הנבחר."
 
-#: anyway/widgets/suburban_widgets/accidents_heat_map_widget.py:52
+#: anyway/widgets/suburban_widgets/accidents_heat_map_widget.py:53
 msgid "Fatal and severe accidents heat map"
 msgstr "מוקדי תאונות קטלניות וקשות במקטע"
 
-#: anyway/widgets/suburban_widgets/head_on_collisions_comparison_widget.py:95
+#: anyway/widgets/suburban_widgets/head_on_collisions_comparison_widget.py:98
 msgid "Fatal head on collisions vs other accidents"
 msgstr "תאונות קטלניות בהתנגשות חזיתית כנגד שאר התאונות"
 
-#: anyway/widgets/suburban_widgets/head_on_collisions_comparison_widget.py:99
+#: anyway/widgets/suburban_widgets/head_on_collisions_comparison_widget.py:102
 msgid "desc"
 msgstr "תאור"
 
-#: anyway/widgets/suburban_widgets/head_on_collisions_comparison_widget.py:126
+#: anyway/widgets/suburban_widgets/head_on_collisions_comparison_widget.py:129
 msgid "others"
 msgstr "אחרות"
 
-#: anyway/widgets/suburban_widgets/head_on_collisions_comparison_widget.py:127
+#: anyway/widgets/suburban_widgets/head_on_collisions_comparison_widget.py:130
 msgid "frontal"
 msgstr "חזיתיות"
 
-#: anyway/widgets/suburban_widgets/head_on_collisions_comparison_widget.py:128
+#: anyway/widgets/suburban_widgets/head_on_collisions_comparison_widget.py:131
 msgid ""
 "Fatal accidents distribution by accident type - head on collisions vs "
 "other accidents."
@@ -364,33 +364,33 @@ msgstr ""
 msgid "Number of Injuries in accidents by severity"
 msgstr "מספר תאונות לפי חומרה"
 
-#: anyway/widgets/suburban_widgets/killed_and_injured_count_per_age_group_stacked_widget.py:52
+#: anyway/widgets/suburban_widgets/killed_and_injured_count_per_age_group_stacked_widget.py:59
 msgid "Killed and injury stacked per age group"
 msgstr "חומרת פגיעה לפי קבוצת גיל"
 
-#: anyway/widgets/suburban_widgets/killed_and_injured_count_per_age_group_stacked_widget.py:53
-#: anyway/widgets/suburban_widgets/killed_and_injured_count_per_age_group_widget.py:38
+#: anyway/widgets/suburban_widgets/killed_and_injured_count_per_age_group_stacked_widget.py:60
+#: anyway/widgets/suburban_widgets/killed_and_injured_count_per_age_group_widget.py:40
 msgid "In"
 msgstr "במקטע"
 
-#: anyway/widgets/suburban_widgets/killed_and_injured_count_per_age_group_widget.py:37
+#: anyway/widgets/suburban_widgets/killed_and_injured_count_per_age_group_widget.py:39
 msgid "Injury per age group"
 msgstr "נפגעים לפי קבוצת גיל"
 
-#: anyway/widgets/suburban_widgets/most_severe_accidents_table_widget.py:36
+#: anyway/widgets/suburban_widgets/most_severe_accidents_table_widget.py:38
 msgid "Most severe accidents in segment"
 msgstr "תאונות חמורות במקטע"
 
-#: anyway/widgets/suburban_widgets/most_severe_accidents_table_widget.py:130
+#: anyway/widgets/suburban_widgets/most_severe_accidents_table_widget.py:132
 msgid "accident_severity"
 msgstr "חומרת תאונה"
 
-#: anyway/widgets/suburban_widgets/most_severe_accidents_table_widget.py:131
+#: anyway/widgets/suburban_widgets/most_severe_accidents_table_widget.py:133
 msgid "type"
 msgstr "סוג"
 
-#: anyway/widgets/suburban_widgets/most_severe_accidents_table_widget.py:143
-#: anyway/widgets/suburban_widgets/most_severe_accidents_table_widget.py:146
+#: anyway/widgets/suburban_widgets/most_severe_accidents_table_widget.py:145
+#: anyway/widgets/suburban_widgets/most_severe_accidents_table_widget.py:148
 msgid ""
 "Most recent fatal and severe accidents in location, ordered by date. Up "
 "to 10 accidents are presented."
@@ -398,7 +398,7 @@ msgstr ""
 "התאונות החמורות (קשות וקטלניות) האחרונות, מסודרות לפי תאריך בסדר יורד. "
 "מוצגות עד 10 תאונות."
 
-#: anyway/widgets/suburban_widgets/most_severe_accidents_widget.py:63
+#: anyway/widgets/suburban_widgets/most_severe_accidents_widget.py:64
 msgid ""
 "Most recent fatal and severe accidents displayed on a map. Up to 10 "
 "accidents are presented."
@@ -406,15 +406,15 @@ msgstr ""
 "התאונות החמורות (קשות וקטלניות) האחרונות מוצגות על גבי מפה. מוצגות עד 10 "
 "תאונות."
 
-#: anyway/widgets/suburban_widgets/motorcycle_accidents_vs_all_accidents_widget.py:137
+#: anyway/widgets/suburban_widgets/motorcycle_accidents_vs_all_accidents_widget.py:138
 msgid "Number of fatal and severe motorcycle accidents"
 msgstr "תאונות אופנוע קטלניות"
 
-#: anyway/widgets/suburban_widgets/motorcycle_accidents_vs_all_accidents_widget.py:137
+#: anyway/widgets/suburban_widgets/motorcycle_accidents_vs_all_accidents_widget.py:138
 msgid "compared to rest of country"
 msgstr "בהשוואה לשאר הארץ"
 
-#: anyway/widgets/suburban_widgets/road2_plus1_widget.py:69
+#: anyway/widgets/suburban_widgets/road2_plus1_widget.py:70
 msgid "Road 2 plus 1 solution to prevent fatal accidents"
 msgstr "פתרון כביש 2 פלוס 1 למניעת תאונות קטלניות"
 
@@ -440,7 +440,7 @@ msgstr "תאונות קטלניות וקשות לכל קילומטר לפי מק
 msgid "Segments with most accidents"
 msgstr "מקטעים עם הכי הרבה תאונות"
 
-#: anyway/widgets/urban_widgets/injured_accidents_with_pedestrians_widget.py:107
+#: anyway/widgets/urban_widgets/injured_accidents_with_pedestrians_widget.py:108
 msgid "Pedestrian accidents by severity and year"
 msgstr "תאונות הולכי רגל לפי חומרה ולפי שנה"
